plotting of errors in XVG class improved
Errors are plotted as bin-percentiles by default. This gives a reasonable
representation of the range of errors. Alternatively, the RMS of the
errors in a bin can be used but that value depends stronger on the
level of coarse graining.

Added additional *_histogrammed_function() to numkit.timeseries, in
particular percentile_histogrammed_function() and additional docs.