canned commands in cbook: shield with try/except
Wrapped the instance generations into a function so that they can be easily caught with
a tr block. Otherwise scripts fail when Gromacs tools are not available and not in all
cases are they necessary to use GW (eg when just using the formats).

Arguably, it would be better to not require cbook in the top level __init__.py ... but
I have not really evaluated if this would break anything.