fixed logging level: each command is logged again
Each command invocation is logged to the log file at level DEBUG (if the logger has been enabled, eg
with gromacs.start_logger()). This did not work for a while because the code for throttling looging
was buggy. With the help of http://stackoverflow.com/questions/6948190/python-logging-reverse-effects-of-disable
it seems to be working now.