xpm: comment on possible fix by Joao M. Damas
On 2014-10-24, Joao mentioned on the gmx-user mailing list that Tsjerks cpm reader
code (on which xpm.py is built) has a bug that is only
noticeable if one is using higher -nlevels for the .xpm construction (in
g_rms, for example). I did NOT incorporate the fix (as it would need testing)
but added a long comment in place with the potential fix.