Added support for multi-index commands
Had to remove GromacsGMXCommand so the gmx 5 commands
would be in the usual OO hierarchy. Also noticed
g_dist is now distance in 5.1.2. Unsure if that was
done at the 5.0 change or 5.1. Could create more problems.
I also noticed gmxcheck is now check. This should solve
the mindist problem in Issue #48